refactor: adjust weex branch coverage and build for merging into dev
