+++ test/unit/karma.cover.config.js
@@ -24,7 +24,8 @@ module.exports = function (config) {
         'test/',
         'src/compiler/parser/html-parser.js',
         'src/core/instance/proxy.js',
-        'src/sfc/deindent.js'
+        'src/sfc/deindent.js',
+        'src/platforms/weex/'
       ]
     }]]
   }
